% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacommunity_variability.r
\name{metacommunity_variability}
\alias{metacommunity_variability}
\title{Metacommunity variability}
\usage{
metacommunity_variability(data_long = NULL, data_wide = NULL,
  time_step_col_name = NULL, site_id_col_name = NULL,
  taxon_list = NULL, taxon_id_col_name = NULL,
  biomass_col_name = NULL, standardization_method = NULL,
  variability_type = NULL)
}
\arguments{
\item{data_long}{Data input for analysis (should be \code{NULL} if input is wide format data). Long format data frame (or tibble) with species counts (or biomass measurements) for multiple sites observed over multiple time periods}

\item{data_wide}{Data input for analysis (should be \code{NULL} if input is long format data). Wide format data frame (or tibble) where rows are observations and columns include \code{time_step_col_name}, \code{site_id_col_name}, taxa listed in \code{taxon_list}.}

\item{time_step_col_name}{Character string indicating time step column name. For use with both \code{data_long} and \code{data_wide} inputs.}

\item{site_id_col_name}{Character string indicating site id column name. For use with both \code{data_long} and \code{data_wide} inputs.}

\item{taxon_list}{Vector (character) of taxon IDs, used to indicate which columns in \code{data_wide} are species counts (or biomass measurements)}

\item{taxon_id_col_name}{Character string indicating column name in \code{data_long} that contains taxon IDs.}

\item{biomass_col_name}{Character string indicating column name in \code{data_long} that contains biomass measurements (or species counts).}

\item{standardization_method}{Only used when \code{variability_type} is "comp". Character string ("h" or "hT") denoting standardization method, either Hellinger ("h") or modified Hellinger ("hT")}

\item{variability_type}{Character string indicating type of analysis. Either "comp" or "agg".}
}
\value{
Returns a one-row data frame where each element names indicate the variability metric that is being returned.
}
\description{
Calculates local (alpha) and regional (gamma) metacommunity variability and the scaling factor phi
}
\examples{
set.seed(123)
dat_example_long <- data.frame(
  my_sites = c(rep('x', 8), rep('y',8)),
  my_times = rep(1:4, each = 2),
  my_spp = rep(c('a','b'), 4),
  my_biomass = (exp(rnorm(16))*10) \%>\% round(0))

# Variability in community composition using the
# modified Hellinger metric
metacommunity_variability(
  data_long = dat_example_long,
  time_step_col_name = 'my_times',
  site_id_col_name = 'my_sites',
  taxon_id_col_name = 'my_spp',
  biomass_col_name = 'my_biomass',
  standardization_method = 'hT', #or 'h'
  variability_type = 'comm')

# Variability in aggregate biomass folloing
# Wang and Loreau (2014)
metacommunity_variability(
  data_long = dat_example_long,
  time_step_col_name = 'my_times',
  site_id_col_name = 'my_sites',
  taxon_id_col_name = 'my_spp',
  biomass_col_name = 'my_biomass',
  variability_type = 'agg')


}
\references{
Wang, S., and M. Loreau. 2014. Ecosystem stability in space: alpha, beta and gamma variability. Ecology Letters 17:891–901. https://doi.org/10.1111/ele.12292.
}
\author{
Eric R. Sokol \email{sokole@gmail.com}

Thomas Lamy \email{thomas.lamy27@gmail.com}
}
